Initialize server modules in the server constructor
This fixes the crash on exit and now the server actually loads a ruleset!
